requirements.txt
openai>=1.0.0
azure-ai-inference>=1.0.0b6
azure-ai-contentsafety>=1.0.0
azure-ai-textanalytics>=5.3.0
azure-search-documents>=11.4.0
azure-core>=1.30.0
python-dotenv>=1.0.0
pytest>=8.0
pytest-timeout>=2.2
